Prevent HTTP 'Server' response header being sent.
For security reasons, production apps ideally shouldn't reveal the
server version that's installed. Unfortunately, thin server seems to
want to announce its presence unless you overwrite the header. This sets
the header to a single space, which appears to be trimmed down to empty
string and subsequently the header is not sent.